Temporarily removed rplot method to make some visualization easier
I need to fix this feature in the future if I want to have all of
my visualizations readily available. This is a low priority task
since I don't need the old visualization I commented out, but this
is a pretty hacky fix for the general case I was trying to support.